+++ packages/core/src/index.ts
@@ -1,6 +1,4 @@
 export * from './Bootstrapable';
-export * from './Configurable';
-export * from './Configure';
 export * from './Context';
 export * from './Environment';
 export * from './Injection';